third_party/blink/web_tests/virtual/unbounded-element-disabled/wpt_internal/html/rendering/unbounded_element/basic-behavior-expected.txt

This is a testharness.js-based test.
[FAIL] Verify the unbounded IDL reflection.
  assert_true: The unbounded IDL attribute should be exposed on HTMLElement. expected true got false
[FAIL] Verify the default user-agent styling for [unbounded].
  assert_equals: Element with unbounded attribute should default to visibility: hidden. expected "hidden" but got "visible"
Harness: the test ran to completion.
